@@ -1,13 +1,14 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 31
+    compileSdkVersion 32
     defaultConfig {
+
         applicationId "com.safesoft.proapp.distribute"
         minSdkVersion 24
-        targetSdkVersion 31
+        targetSdkVersion 32
         versionCode 537
-        versionName "1.2"
+        versionName "1.3"
         testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
         multiDexEnabled true //important
 
@@ -60,10 +61,10 @@ dependencies {
 
     implementation 'com.haozhang.libary:android-slanted-textview:1.2'
     implementation 'com.haozhang.libary:android-slanted-textview:1.2'
-    implementation 'androidx.appcompat:appcompat:1.4.1'
+    implementation 'androidx.appcompat:appcompat:1.4.2'
     implementation 'androidx.cardview:cardview:1.0.0'
-    implementation 'com.google.android.material:material:1.5.0'
-    implementation 'androidx.constraintlayout:constraintlayout:2.1.3'
+    implementation 'com.google.android.material:material:1.6.1'
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     implementation 'com.amulyakhare:com.amulyakhare.textdrawable:1.0.1'
     implementation 'org.greenrobot:eventbus:3.0.0'
     implementation 'com.uniquestudio:checkbox:1.0.1'
@@ -74,7 +75,7 @@ dependencies {
     implementation 'com.github.paolorotolo:expandableheightlistview:1.0.0'
     implementation 'com.victor:lib:1.0.4'
     implementation 'eu.inloop:simplerecycleradapter:0.3.1'
-    implementation 'me.grantland:autofittextview:0.2.+'
+    implementation 'me.grantland:autofittextview:0.2.1'
     implementation 'com.jzxiang.pickerview:TimePickerDialog:1.0.1'
     implementation 'com.weiwangcn.betterspinner:library-material:1.1.0'
     implementation 'com.jakewharton:butterknife:8.0.1'
@@ -95,12 +96,17 @@ dependencies {
     implementation 'com.github.florent37:singledateandtimepicker:1.2.2'
    // implementation 'com.github.nekocode:Badge:{lastest-version}'
 
+    implementation 'com.github.medyo:fancybuttons:1.9.1'
+
+    implementation 'com.google.maps:google-maps-services:2.1.0'
+    implementation 'org.slf4j:slf4j-nop:1.7.25'
+
 
 
     // externel librery
     implementation project(':folding-cell')
     implementation project(':lib_badge')
-    implementation project(':library_bth')
+   // implementation project(':library_bth')
     implementation project(':materialbarcodescanner')
     implementation project(path: ':SpinKitLib')
     implementation project(':sweet_alert_dialog_library')
@@ -112,7 +118,9 @@ dependencies {
     implementation files('libs/httpmime-4.0.3.jar')
     implementation files('libs/dropbox-android-sdk-1.6.3.jar')
     testImplementation 'junit:junit:4.13.2'
-    implementation files('libs/commons-net-1.4.1.jar')
+    implementation files('libs/commons-net-3.8.0.jar')
+    implementation files('libs/printer_library.jar')
+    //implementation project(path: ':libezftp')
 
     implementation ("com.journeyapps:zxing-android-embedded:4.2.0"){
         exclude group: 'com.google.zxing', module: 'android-core'
