@@ -1,21 +1,25 @@
 package com.safesoft.proapp.distribute.adapters;
 
+import android.annotation.SuppressLint;
 import android.content.Context;
 import android.graphics.Color;
 import android.graphics.Typeface;
+
+import androidx.annotation.NonNull;
 import androidx.cardview.widget.CardView;
 import androidx.recyclerview.widget.RecyclerView;
 import android.text.SpannableString;
 import android.text.TextUtils;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.LinearLayout;
 import android.widget.TextView;
 
 import com.haozhang.lib.SlantedTextView;
 import com.safesoft.proapp.distribute.R;
 import com.safesoft.proapp.distribute.postData.PostData_Bon1;
-import com.safesoft.proapp.distribute.util.ColorGeneratorModified;
-import com.safesoft.proapp.distribute.util.MyCardView2;
+import com.safesoft.proapp.distribute.utils.ColorGeneratorModified;
+import com.safesoft.proapp.distribute.utils.MyCardView2;
 
 import java.text.DecimalFormat;
 import java.util.List;
@@ -28,45 +32,52 @@ import cn.nekocode.badge.BadgeDrawable;
 
 public class RecyclerAdapterBon1 extends RecyclerView.Adapter<RecyclerAdapterBon1.MyViewHolder> {
 
-    private List<PostData_Bon1> bon1List;
+    private final List<PostData_Bon1> bon1List;
     private int color = 0;
     private ItemClick itemClick;
     private ItemLongClick itemLongClick;
     private ColorGeneratorModified generator;
-    private Context mContext;
+    private String SOURCE;
 
 
-    class MyViewHolder extends RecyclerView.ViewHolder {
+    static class MyViewHolder extends RecyclerView.ViewHolder {
 
         TextView NumBon;
         TextView NomClient;
         TextView Montant;
         TextView nbrProduit;
         TextView Date_bon;
+        TextView Heure_bon;
+        TextView Versement;
         CardView cardView;
         SlantedTextView blocage;
+        LinearLayout lnr_versement;
 
         MyViewHolder(View view) {
             super(view);
 
-            cardView = (CardView) view.findViewById(R.id.item_root);
-            NumBon = (TextView) view.findViewById(R.id.num_bon);
-            NomClient = (TextView) view.findViewById(R.id.nom_client);
-            Montant = (TextView) view.findViewById(R.id.montant);
-            nbrProduit = (TextView) view.findViewById(R.id.nbr_p);
-            Date_bon = (TextView) view.findViewById(R.id.date_bon);
-            blocage = (SlantedTextView) view.findViewById(R.id.blocage);
+            cardView = view.findViewById(R.id.item_root);
+            NumBon = view.findViewById(R.id.num_bon);
+            NomClient = view.findViewById(R.id.nom_client);
+            Montant = view.findViewById(R.id.montant);
+            nbrProduit = view.findViewById(R.id.nbr_p);
+            Date_bon = view.findViewById(R.id.date_bon);
+            Heure_bon = view.findViewById(R.id.heure_bon);
+            Versement = view.findViewById(R.id.versement_bon);
+            blocage = view.findViewById(R.id.blocage);
+            lnr_versement = view.findViewById(R.id.lnr_versement);
         }
     }
 
 
-    public RecyclerAdapterBon1(Context context, List<PostData_Bon1> itemList) {
+    public RecyclerAdapterBon1(Context context, List<PostData_Bon1> itemList, String SOURCE) {
         this.bon1List = itemList;
+        this.SOURCE =SOURCE;
         if (color == 0)
             generator = ColorGeneratorModified.MATERIAL;
-        mContext = context;
     }
 
+    @NonNull
     @Override
     public MyViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
         View v = new MyCardView2(parent.getContext(), R.layout.item_bon1_list);
@@ -77,6 +88,7 @@ public class RecyclerAdapterBon1 extends RecyclerView.Adapter<RecyclerAdapterBon
         return new MyViewHolder(v);
     }
 
+    @SuppressLint("SetTextI18n")
     @Override
     public void onBindViewHolder(final MyViewHolder holder,int position) {
         PostData_Bon1 item = bon1List.get(position);
@@ -88,19 +100,30 @@ public class RecyclerAdapterBon1 extends RecyclerView.Adapter<RecyclerAdapterBon
         holder.NomClient.setText(""+item.client);
 
         if(item.montant_bon == null){
-            item.montant_bon = "0.00";
+            item.montant_bon = 0.00;
         }
-        holder.Montant.setText(""+ new DecimalFormat("##,##0.00").format(Double.valueOf(item.montant_bon)) + " DA");
+        if(item.verser == null){
+            item.verser = 0.00;
+        }
+
+        if(SOURCE.equals("SALE")){
+            holder.lnr_versement.setVisibility(View.VISIBLE);
+            holder.Versement.setText(""+ new DecimalFormat("##,##0.00").format(item.verser) + " DA");
+        }else{
+            holder.lnr_versement.setVisibility(View.GONE);
+        }
+
+        holder.Montant.setText(""+ new DecimalFormat("##,##0.00").format(item.montant_bon) + " DA");
 
         if(item.nbr_p == null)
         {
-            item.nbr_p = "0";
+            item.nbr_p = 0;
         }
         final BadgeDrawable drawable1 = new BadgeDrawable.Builder()
                 .type(BadgeDrawable.TYPE_NUMBER)
                 .badgeColor(0xff303F9F)
                 .textSize(35)
-                .number(Integer.valueOf(item.nbr_p))
+                .number(item.nbr_p)
                 .build();
 
 
@@ -110,34 +133,29 @@ public class RecyclerAdapterBon1 extends RecyclerView.Adapter<RecyclerAdapterBon
 
         holder.Date_bon.setText(item.date_bon);
 
-        holder.cardView.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                itemClick.onClick(view,holder.getAdapterPosition());
-            }
+        holder.Heure_bon.setText(item.heure);
+
+        holder.cardView.setOnClickListener(view -> itemClick.onClick(view,holder.getAdapterPosition()));
+
+        holder.cardView.setOnLongClickListener(v -> {
+            itemLongClick.onLongClick(v , holder.getAdapterPosition());
+            return true;
         });
 
-        holder.cardView.setOnLongClickListener(new View.OnLongClickListener() {
-            @Override
-            public boolean onLongClick(View v) {
-                itemLongClick.onLongClick(v , holder.getAdapterPosition());
-                return true;
-            }
-        });
 
         if(item.blocage.equals("F")){
             holder.blocage.setText("Validé")
                     .setTextColor(Color.WHITE)
                     .setSlantedBackgroundColor(Color.GREEN)
                     .setTextSize(21)
-                    .setSlantedLength(50)
+                    .setSlantedLength(80)
                     .setMode(SlantedTextView.MODE_RIGHT_BOTTOM);
         }else {
             holder.blocage.setText("En attente")
                     .setTextColor(Color.WHITE)
                     .setSlantedBackgroundColor(Color.RED)
                     .setTextSize(21)
-                    .setSlantedLength(50)
+                    .setSlantedLength(80)
                     .setMode(SlantedTextView.MODE_RIGHT_BOTTOM);
         }
 
@@ -161,6 +179,7 @@ public class RecyclerAdapterBon1 extends RecyclerView.Adapter<RecyclerAdapterBon
         void onLongClick(View v, int position);
     }
 
+    @SuppressLint("NotifyDataSetChanged")
     public void refresh(List<PostData_Bon1> new_itemList){
         bon1List.clear();
         bon1List.addAll(new_itemList);
