@@ -59,9 +59,6 @@ import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Date;
 
-import app.akexorcist.bluetotohspp.library.BluetoothSPP;
-import app.akexorcist.bluetotohspp.library.BluetoothState;
-import app.akexorcist.bluetotohspp.library.DeviceList;
 import cn.pedant.SweetAlert.SweetAlertDialog;
 import de.keyboardsurfer.android.widget.crouton.Crouton;
 import de.keyboardsurfer.android.widget.crouton.Style;
@@ -72,31 +69,7 @@ import eu.inloop.simplerecycleradapter.SimpleRecyclerAdapter;
 
 public class ActivityEtatV extends AppCompatActivity implements ItemClickListener<WrappedMyDataObject>, ItemLongClickListener<WrappedMyDataObject> {
 
-  ////////////////////////////////////////
-  private final int NOPAPER = 3;
-  private final int LOWBATTERY = 4;
-  private final int PRINTVERSION = 5;
-  private final int PRINTBARCODE = 6;
-  private final int PRINTQRCODE = 7;
-  private final int PRINTPAPERWALK = 8;
-  private final int PRINTCONTENT = 9;
-  private final int CANCELPROMPT = 10;
-  private final int PRINTERR = 11;
-  private final int OVERHEAT = 12;
-  private final int MAKER = 13;
-  private final int PRINTPICTURE = 14;
-  private final int EXECUTECOMMAND = 15;
-
-  private String Result;
-  private Boolean nopaper = false;
-  private boolean LowBattery = false;
-  private ProgressDialog progressDialog;
-  private ProgressDialog progressDialog_wait_connecte;
-  private final static int MAX_LEFT_DISTANCE = 255;
-  private ProgressDialog dialog;
-  private MyHandler_Integrate handler_Integrate;
-  private MyHandler_Bluetooth handler_Bluetooth;
-  private String PREFS_PRINTER = "ConfigPrinter";
+  private String PREFS = "ALL_PREFS";
 
   private RecyclerView mRecyclerView;
   private SimpleRecyclerAdapter<WrappedMyDataObject> mRecyclerAdapter;
@@ -108,8 +81,7 @@ public class ActivityEtatV extends AppCompatActivity implements ItemClickListene
   private RelativeLayout empty_data;
   private TextView debut, fin,user;
   private ViewGroup mainView;
-  String clientt;
-  private BluetoothSPP bt;
+
   DATABASE controller;
   private MediaPlayer mp;
   Boolean printer_mode_integrate = true;
@@ -119,14 +91,12 @@ public class ActivityEtatV extends AppCompatActivity implements ItemClickListene
 
   private ArrayList<PostData_Etatv> result_etatzg;
   private EventBus bus = EventBus.getDefault();
-  private String from_d;
+
   private String client;
   private String c_client;
 
-
+  private String from_d;
   private String to_d;
-  private String from_h;
-  private String to_h;
 
   @Override
   protected void onCreate(Bundle savedInstanceState) {
@@ -136,47 +106,6 @@ public class ActivityEtatV extends AppCompatActivity implements ItemClickListene
     // Register as a subscriber
     bus.register(this);
 
-    IntentFilter pIntentFilter = new IntentFilter();
-    pIntentFilter.addAction(Intent.ACTION_BATTERY_CHANGED);
-    pIntentFilter.addAction("android.intent.action.BATTERY_CAPACITY_EVENT");
-    registerReceiver(printReceive, pIntentFilter);
-
-
-    bt = new BluetoothSPP(this);
-
-    if(!bt.isBluetoothAvailable()) {
-      Toast.makeText(getApplicationContext(), "Bluetooth is not available", Toast.LENGTH_SHORT).show();
-      finish();
-    }
-
-    bt.setBluetoothConnectionListener(new BluetoothSPP.BluetoothConnectionListener() {
-
-      public void onDeviceDisconnected() {
-        if (progressDialog_wait_connecte != null && !ActivityEtatV.this.isFinishing()) {
-          progressDialog_wait_connecte.dismiss();
-          progressDialog_wait_connecte = null;
-        }
-        Crouton.makeText(ActivityEtatV.this, "Imprimente bluetooth non connecté ", Style.ALERT).show();
-      }
-
-      public void onDeviceConnectionFailed() {
-        if (progressDialog_wait_connecte != null && !ActivityEtatV.this.isFinishing()) {
-          progressDialog_wait_connecte.dismiss();
-          progressDialog_wait_connecte = null;
-        }
-        Crouton.makeText(ActivityEtatV.this, "Imprimente connection erroné ", Style.ALERT).show();
-      }
-
-      public void onDeviceConnected(String name, String address) {
-        if (progressDialog_wait_connecte != null && !ActivityEtatV.this.isFinishing()) {
-          progressDialog_wait_connecte.dismiss();
-          progressDialog_wait_connecte = null;
-        }
-        Crouton.makeText(ActivityEtatV.this, "Imprimente connecté à  "+ name, Style.CONFIRM).show();
-        prepareBon_Bluetooth();
-      }
-
-    });
   }
 
   @Override
@@ -186,10 +115,7 @@ public class ActivityEtatV extends AppCompatActivity implements ItemClickListene
 
     getSupportActionBar().setDisplayHomeAsUpEnabled(true);
     getSupportActionBar().setTitle("Statistique des ventes");
-    getSupportActionBar().setBackgroundDrawable(new ColorDrawable(getResources()
-            .getColor(R.color.black)));
     empty_data.setVisibility(View.VISIBLE);
-
     retry =(ImageView) findViewById(R.id.retry);
     retry.setOnClickListener(new View.OnClickListener() {
       @Override
@@ -205,24 +131,11 @@ public class ActivityEtatV extends AppCompatActivity implements ItemClickListene
 
     initAdapter();
 
-    SharedPreferences prefs1 = getSharedPreferences(PREFS_PRINTER, MODE_PRIVATE);
-    if(prefs1.getString("PRINTER", "INTEGRATE").toString().equals("INTEGRATE")){
+    SharedPreferences prefs1 = getSharedPreferences(PREFS, MODE_PRIVATE);
+    if(prefs1.getString("PRINTER", "INTEGRATE").equals("INTEGRATE")){
       printer_mode_integrate = true;
     }else {
       printer_mode_integrate = false;
-    }
-
-    if(!printer_mode_integrate){
-      if (!bt.isBluetoothEnabled()) {
-        Intent intent = new Intent(BluetoothAdapter.ACTION_REQUEST_ENABLE);
-        startActivityForResult(intent, BluetoothState.REQUEST_ENABLE_BT);
-      } else {
-        if(!bt.isServiceAvailable()) {
-          bt.setupService();
-          bt.startService(BluetoothState.DEVICE_ANDROID);
-          setup();
-        }
-      }
     }
 
     super.onStart();
@@ -264,41 +177,18 @@ public class ActivityEtatV extends AppCompatActivity implements ItemClickListene
     event_selection = event;
     debut.setText("De "+ event_selection.getDate_f());
     fin.setText("Vers "+ event_selection.getDate_t());
-    if(event.getUser().toString().equals("%")){
+    if(event.getUser().equals("%")){
       user.setText("Tous");
     }else{
       user.setText(" "+ event_selection.getUser());
     }
     client = event_selection.getUser();
     c_client = event_selection.getCode_user();
-    SimpleDateFormat df_show = new SimpleDateFormat("dd/MM/yyyy");
-
-    SimpleDateFormat df_save = new SimpleDateFormat("MM/dd/yyyy");
-
-    from_d = event_selection.getDate_f().substring(0, event_selection.getDate_f().indexOf(" "));
-
-    Date myDate = null;
-    try {
-      myDate = df_save.parse(from_d);
-
-    } catch (ParseException e) {
-      e.printStackTrace();
-    }
-    from_d = df_show.format(myDate);
 
 
-    to_d = event_selection.getDate_t().substring(0, event_selection.getDate_t().indexOf(" "));;
-    Date myDate2 = null;
-    try {
-      myDate2 = df_save.parse(to_d);
+    from_d = event_selection.getDate_f();
+    to_d = event_selection.getDate_t();
 
-    } catch (ParseException e) {
-      e.printStackTrace();
-    }
-    to_d = df_show.format(myDate2);
-
-    from_h =  event_selection.getDate_f().substring(event_selection.getDate_f().indexOf(" ") + 1, event_selection.getDate_f().length());
-    to_h = event_selection.getDate_t().substring(event_selection.getDate_t().indexOf(" ") + 1, event_selection.getDate_t().length());
     get_etatzg();
   }
 
@@ -352,15 +242,15 @@ public class ActivityEtatV extends AppCompatActivity implements ItemClickListene
     //  mRecyclerAdapter.addItem(WrappedMyDataObject.initHeaderItem(result_etat_z.get(i).produit, result_etat_z.get(i).quantite, result_etat_z.get(i).montant));
 
     for (int i = 0; i < result_etat_z.size(); i++) {
-      if (result_etat_z.get(i).code_parent.toString().equals("1")) {
+      if (result_etat_z.get(i).code_parent.equals("1")) {
         mRecyclerAdapter.addItem(WrappedMyDataObject.initDataItem(new MyDataObject(result_etat_z.get(i).produit, result_etat_z.get(i).quantite, result_etat_z.get(i).montant)));
-      } else if (result_etat_z.get(i).code_parent.toString().equals("-6")) {
+      } else if (result_etat_z.get(i).code_parent.equals("-6")) {
         mRecyclerAdapter.addItem(WrappedMyDataObject.initHeaderItemTotal("Conclusion Total : "));
         for (int k = i; k < result_etat_z.size() - 1; k++) {
           mRecyclerAdapter.addItem(WrappedMyDataObject.initDataItem(new MyDataObject(result_etat_z.get(k).produit, result_etat_z.get(k).quantite, result_etat_z.get(k).montant)));
           i = k;
         }
-      } else if (result_etat_z.get(i).code_parent.toString().equals("-8")) {
+      } else if (result_etat_z.get(i).code_parent.equals("-8")) {
         mRecyclerAdapter.addItem(WrappedMyDataObject.initHeaderItemTotal("Objectif : "));
         mRecyclerAdapter.addItem(WrappedMyDataObject.initDataItemObjectif(new MyDataObject(result_etat_z.get(i).produit, result_etat_z.get(i).quantite, result_etat_z.get(i).montant)));
       }
@@ -460,18 +350,7 @@ public class ActivityEtatV extends AppCompatActivity implements ItemClickListene
       case R.id.print:
         if(result_etatzg.size() > 0){
           if(printer_mode_integrate){
-            prepareBon_IntegratePrinter();
-          }else{
-            if(bt.getServiceState() == BluetoothState.STATE_CONNECTED){
-              // print
-              Toast.makeText(ActivityEtatV.this, " Impression ..." , Toast.LENGTH_SHORT).show();
-              prepareBon_Bluetooth();
 
-            }else{
-              bt.setDeviceTarget(BluetoothState.DEVICE_OTHER);
-              Intent intent = new Intent(getApplicationContext(), DeviceList.class);
-              startActivityForResult(intent, BluetoothState.REQUEST_CONNECT_DEVICE);
-            }
           }
         }
 
@@ -532,18 +411,13 @@ public class ActivityEtatV extends AppCompatActivity implements ItemClickListene
           if(result_etatzg != null ) {
             result_etatzg.clear();
           }
-if(c_client == null)
-{
-  flag = getEtatGlobal(from_d,  to_d,  from_h,  to_h);
-
-
-}else
-{
-  flag = getEtatzgs(c_client,from_d,  to_d,  from_h,  to_h);
-
-}
-
-
+          if(c_client == null)
+          {
+            flag = getEtatGlobal(from_d,  to_d);
+          }else
+          {
+            flag = getEtatzgs(c_client,from_d,  to_d);
+          }
           if (flag == 0) {
             //failed
             handler.sendEmptyMessage(3);
@@ -564,11 +438,11 @@ if(c_client == null)
     thread.start();
   }
 
-  public int getEtatzgs(String c_client,String from_d, String to_d, String from_h, String to_h){
+  public int getEtatzgs(String c_client,String from_d, String to_d){
     int flag = 0;
     try {
 
-      result_etatzg =  controller.select_etatv_from_database(c_client,from_d,  to_d,  from_h,  to_h);
+      result_etatzg =  controller.select_etatv_from_database(c_client,from_d,  to_d);
       flag = 1;
     }catch (Exception sqle){
       Log.v("TRACKKK", sqle.getMessage());
@@ -576,11 +450,10 @@ if(c_client == null)
     }
     return flag;
   }
-  public int getEtatGlobal(String from_d, String to_d, String from_h, String to_h){
+  public int getEtatGlobal(String from_d, String to_d){
     int flag = 0;
     try {
-
-      result_etatzg =  controller.select_etatv_global_from_database(from_d,  to_d,  from_h,  to_h);
+      result_etatzg =  controller.select_etatv_global_from_database(from_d,  to_d);
       flag = 1;
     }catch (Exception sqle){
       Log.v("TRACKKK", sqle.getMessage());
@@ -603,569 +476,10 @@ if(c_client == null)
 
   @Override
   protected void onDestroy() {
-    // Unregister
-    if (progressDialog != null && !ActivityEtatV.this.isFinishing()) {
-      progressDialog.dismiss();
-      progressDialog = null;
-    }
-    unregisterReceiver(printReceive);
-    ThermalPrinter.stop();
-
-    if(bt != null)
-      bt.stopService();
 
     bus.unregister(this);
+
     super.onDestroy();
   }
 
-
-  // Preparation de Bon pour l'imprimente intégrer
-  protected void prepareBon_IntegratePrinter(){
-
-    handler_Integrate = new ActivityEtatV.MyHandler_Integrate();
-    if (LowBattery == true) {
-      handler_Integrate.sendMessage(handler_Integrate.obtainMessage(LOWBATTERY, 1, 0, null));
-    } else {
-      if (!nopaper) {
-        progressDialog = ProgressDialog.show(ActivityEtatV.this, getString(R.string.bl_dy), getString(R.string.printing_wait));
-        handler_Integrate.sendMessage(handler_Integrate.obtainMessage(PRINTCONTENT, 1, 0, null));
-      } else {
-        Toast.makeText(ActivityEtatV.this, getString(R.string.ptintInit), Toast.LENGTH_LONG).show();
-      }
-    }
-  }
-
-
-  protected void prepareBon_Bluetooth(){
-
-    handler_Bluetooth = new ActivityEtatV.MyHandler_Bluetooth();
-    if (LowBattery == true) {
-      handler_Bluetooth.sendMessage(handler_Bluetooth.obtainMessage(LOWBATTERY, 1, 0, null));
-    } else {
-      if (!nopaper) {
-        progressDialog = ProgressDialog.show(ActivityEtatV.this, getString(R.string.bl_dy), getString(R.string.printing_wait));
-        handler_Bluetooth.sendMessage(handler_Bluetooth.obtainMessage(PRINTCONTENT, 1, 0, null));
-      } else {
-        Toast.makeText(ActivityEtatV.this, getString(R.string.ptintInit), Toast.LENGTH_LONG).show();
-      }
-    }
-  }
-  //Class thread printing
-  private class contentPrintThread_Integrate extends Thread {
-    @Override
-    public void run() {
-      super.run();
-
-      try {
-        ThermalPrinter.start(ActivityEtatV.this);
-        ThermalPrinter.reset();
-        ThermalPrinter.setAlgin(ThermalPrinter.ALGIN_LEFT);
-
-        /////////////////////////////
-        ThermalPrinter.reset();
-        ThermalPrinter.setLeftIndent(0);
-        ThermalPrinter.setLineSpace(28);
-        ThermalPrinter.setFontSize(2);
-        ThermalPrinter.setGray(12);
-
-
-
-        ThermalPrinter.reset();
-        ThermalPrinter.setLeftIndent(0);
-        ThermalPrinter.setLineSpace(28);
-        ThermalPrinter.setFontSize(2);
-        ThermalPrinter.enlargeFontSize(2, 2);
-        ThermalPrinter.setGray(12);
-
-        ////////////////////////////////
-        //  ThermalPrinter.setBold(true);
-        ThermalPrinter.addString(" Statistique des ventes\n");
-        ThermalPrinter.addString("----------------");
-        ThermalPrinter.printString();
-        /////////////////////////////
-        ThermalPrinter.reset();
-        ThermalPrinter.setLeftIndent(0);
-        ThermalPrinter.setLineSpace(28);
-        ThermalPrinter.setFontSize(2);
-        ThermalPrinter.setGray(12);
-        ////////////////////////////////
-        ThermalPrinter.addString("CLIENT : " + event_selection.getUser());
-        ThermalPrinter.printString();
-        ThermalPrinter.addString("DE : " + event_selection.getDate_f());
-        ThermalPrinter.printString();
-        ThermalPrinter.addString("VERS : " + event_selection.getDate_t());
-        ThermalPrinter.printString();
-        ThermalPrinter.addString("--------------------------------");
-        ThermalPrinter.printString();
-        ///////////////////////
-        ThermalPrinter.reset();
-        ThermalPrinter.setLeftIndent(0);
-        ThermalPrinter.setLineSpace(28);
-        ThermalPrinter.setFontSize(1);
-        ThermalPrinter.enlargeFontSize(1, 2);
-        ThermalPrinter.setGray(12);
-        ///////////////////////////
-
-        for(int b= 0; b<result_etatzg.size(); b++){
-          if(result_etatzg.get(b).code_parent.equals("1")){
-            if(result_etatzg.get(b).produit.length()>48){
-              result_etatzg.get(b).produit = result_etatzg.get(b).produit.substring(0, 47);
-            }
-            ThermalPrinter.addString(result_etatzg.get(b).produit);
-            ThermalPrinter.printString();
-
-            String quantite = result_etatzg.get(b).quantite;
-            String prix_u = new DecimalFormat("##,##0.00").format(Double.valueOf(result_etatzg.get(b).pv_ht));
-            String total_produit = new DecimalFormat("##,##0.00").format(Double.valueOf(Double.valueOf(result_etatzg.get(b).montant)));
-            String espace_walk0="", espace_walk1="", espace_walk2="", espace_walk3= "", espace_walk4= "";
-
-            //Walk 1
-            if(quantite.length() == 1) {
-              espace_walk0 = "     ";
-            }else  if(quantite.length() == 2){
-              espace_walk0 = "    ";
-            }else  if(quantite.length() == 3){
-              espace_walk0 = "   ";
-            }else  if(quantite.length() == 4){
-              espace_walk0 = "  ";
-            }else  if(quantite.length() == 5){
-              espace_walk0 = " ";
-            }
-
-            espace_walk1 = "   ";
-
-            //Walk 2
-            if(prix_u.length() ==4){
-              espace_walk2 = "          ";
-            }else if(prix_u.length() ==5){
-              espace_walk2 = "         ";
-            }else if(prix_u.length() ==6){
-              espace_walk2 = "        ";
-            }else if(prix_u.length() ==7){
-              espace_walk2 = "       ";
-            }else if(prix_u.length() ==8){
-              espace_walk2 = "      ";
-            }else if(prix_u.length() ==9){
-              espace_walk2 = "     ";
-            }else if(prix_u.length() ==10){
-              espace_walk2 = "    ";
-            }else if(prix_u.length() ==11){
-              espace_walk2 = "   ";
-            }else if(prix_u.length() ==12){
-              espace_walk2 = "  ";
-            }
-
-
-            //walk3
-            espace_walk3 = "   ";
-            // walk4
-            if(total_produit.length() ==4){
-              espace_walk4 = "                ";
-            }else if(total_produit.length() ==5){
-              espace_walk4 = "               ";
-            }else if(total_produit.length() ==6){
-              espace_walk4 = "              ";
-            }else if(total_produit.length() ==7){
-              espace_walk4 = "             ";
-            }else if(total_produit.length() ==8){
-              espace_walk4 = "            ";
-            }else if(total_produit.length() ==9){
-              espace_walk4 = "           ";
-            }else if(total_produit.length() ==10){
-              espace_walk4 = "          ";
-            }else if(total_produit.length() ==11){
-              espace_walk4 = "         ";
-            }else if(total_produit.length() ==12){
-              espace_walk4 = "        ";
-            }else if(total_produit.length() ==13){
-              espace_walk4 = "       ";
-            }else if(total_produit.length() ==14){
-              espace_walk4 = "      ";
-            }else if(total_produit.length() ==15){
-              espace_walk4 = "     ";
-            }else if(total_produit.length() ==16){
-              espace_walk4 = "    ";
-            }else if(total_produit.length() ==17){
-              espace_walk4 = "   ";
-            }else if(total_produit.length() ==18){
-              espace_walk4 = "  ";
-            }
-
-            ThermalPrinter.addString(espace_walk0 +quantite+ espace_walk1 +"X"+ espace_walk2 + prix_u + espace_walk3 +"="+ espace_walk4 + total_produit);
-            ThermalPrinter.printString();
-          }
-        }
-
-        ThermalPrinter.addString("------------------------------------------------");
-        ThermalPrinter.printString();
-        /////////////////////////////
-        ThermalPrinter.reset();
-        ThermalPrinter.setLeftIndent(0);
-        ThermalPrinter.setLineSpace(28);
-        ThermalPrinter.setFontSize(2);
-        ThermalPrinter.setGray(12);
-        /////////////////////////////////////////////
-
-        for(int s = 0; s < result_etatzg.size(); s ++){
-
-          if(result_etatzg.get(s).code_parent.equals("-6")) {
-
-            String walk = "";
-
-            if(result_etatzg.get(s).quantite != null){
-              String total_ht = "TOTAL QTE : " + new DecimalFormat("##,##0").format(Double.valueOf(result_etatzg.get(s).quantite));
-              for (int a = 0; a < 32 - total_ht.length(); a++) {
-                walk = walk + " ";
-              }
-              ThermalPrinter.addString(walk + total_ht);
-              ThermalPrinter.printString();
-            }
-
-            if(result_etatzg.get(s).montant != null){
-              walk = "";
-              String tva = "MONTANT TOTAL : " + new DecimalFormat("##,##0.00").format(Double.valueOf(result_etatzg.get(s).montant)) + " DA";
-              for (int a = 0; a < 32 - tva.length(); a++) {
-                walk = walk + " ";
-              }
-              ThermalPrinter.addString(walk + tva);
-              ThermalPrinter.printString();
-            }
-          }
-        }
-
-        ThermalPrinter.walkPaper(100);
-
-      } catch (TelpoException e) {
-        e.printStackTrace();
-        Result = e.toString();
-        if (Result.equals("com.telpo.tps550.api.printer.NoPaperException")) {
-          nopaper = true;
-        } else if (Result.equals("com.telpo.tps550.api.printer.OverHeatException")) {
-          handler_Integrate.sendMessage(handler_Integrate.obtainMessage(OVERHEAT, 1, 0, null));
-        } else {
-          handler_Integrate.sendMessage(handler_Integrate.obtainMessage(PRINTERR, 1, 0, null));
-        }
-      } finally {
-        handler_Integrate.sendMessage(handler_Integrate.obtainMessage(CANCELPROMPT, 1, 0, null));
-        if (nopaper){
-          handler_Integrate.sendMessage(handler_Integrate.obtainMessage(NOPAPER, 1, 0, null));
-          nopaper = false;
-          return;
-        }
-        ThermalPrinter.stop(ActivityEtatV.this);
-      }
-    }
-  }
-
-  private class MyHandler_Integrate extends Handler {
-    @Override
-    public void handleMessage(Message msg) {
-      switch (msg.what) {
-        case NOPAPER:
-          noPaperDlg();
-          break;
-        case LOWBATTERY:
-          AlertDialog.Builder alertDialog = new AlertDialog.Builder(ActivityEtatV.this);
-          alertDialog.setTitle(R.string.operation_result);
-          alertDialog.setMessage(getString(R.string.LowBattery));
-          alertDialog.setPositiveButton(getString(R.string.dialog_comfirm), new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialogInterface, int i) {
-            }
-          });
-          alertDialog.show();
-          break;
-        case PRINTVERSION:
-          dialog.dismiss();
-          if (msg.obj.equals("1")) {
-            // textPrintVersion.setText(printVersion);
-          } else {
-            Toast.makeText(ActivityEtatV.this, "Operation failed", Toast.LENGTH_LONG).show();
-          }
-          break;
-        case PRINTBARCODE:
-          // new barcodePrintThread().start();
-          break;
-        case PRINTQRCODE:
-          //  new qrcodePrintThread().start();
-          break;
-        case PRINTPAPERWALK:
-          //   new paperWalkPrintThread().start();
-          break;
-        case PRINTCONTENT:
-          new contentPrintThread_Integrate().start();
-          break;
-        case MAKER:
-          //   new MakerThread().start();
-          break;
-        case PRINTPICTURE:
-          //   new printPicture().start();
-          break;
-        case CANCELPROMPT:
-          if (progressDialog != null && !ActivityEtatV.this.isFinishing()) {
-            progressDialog.dismiss();
-            progressDialog = null;
-          }
-          break;
-        case EXECUTECOMMAND:
-          //  new executeCommand().start();
-          break;
-        case OVERHEAT:
-          AlertDialog.Builder overHeatDialog = new AlertDialog.Builder(ActivityEtatV.this);
-          overHeatDialog.setTitle(R.string.operation_result);
-          overHeatDialog.setMessage(getString(R.string.overTemp));
-          overHeatDialog.setPositiveButton(getString(R.string.dialog_comfirm), new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialogInterface, int i) {
-            }
-          });
-          overHeatDialog.show();
-          break;
-        default:
-          Toast.makeText(ActivityEtatV.this, "Print Error!", Toast.LENGTH_LONG).show();
-          break;
-      }
-    }
-  }
-
-  private final BroadcastReceiver printReceive = new BroadcastReceiver() {
-    @Override
-    public void onReceive(Context context, Intent intent) {
-      String action = intent.getAction();
-      if (action.equals(Intent.ACTION_BATTERY_CHANGED)) {
-        int status = intent.getIntExtra(BatteryManager.EXTRA_STATUS, BatteryManager.BATTERY_STATUS_NOT_CHARGING);
-        int level = intent.getIntExtra(BatteryManager.EXTRA_LEVEL, 0);
-        int scale = intent.getIntExtra(BatteryManager.EXTRA_SCALE, 0);
-        //TPS390 can not print,while in low battery,whether is charging or not charging
-          if (status != BatteryManager.BATTERY_STATUS_CHARGING) {
-            if (level * 5 <= scale) {
-              LowBattery = true;
-            } else {
-              LowBattery = false;
-            }
-          } else {
-            LowBattery = false;
-          }
-      }
-      //Only use for TPS550MTK devices
-      else if (action.equals("android.intent.action.BATTERY_CAPACITY_EVENT")) {
-        int status = intent.getIntExtra("action", 0);
-        int level = intent.getIntExtra("level", 0);
-        if(status == 0){
-          if(level < 1){
-            LowBattery = true;
-          }else {
-            LowBattery = false;
-          }
-        }else {
-          LowBattery = false;
-        }
-      }
-    }
-  };
-
-  @Override
-  public boolean onKeyDown(int keyCode, KeyEvent event) {
-    if (keyCode == KeyEvent.KEYCODE_BACK) {
-      finish();
-    }
-    return super.onKeyDown(keyCode, event);
-  }
-
-  private void noPaperDlg() {
-    AlertDialog.Builder dlg = new AlertDialog.Builder(ActivityEtatV.this);
-    dlg.setTitle(getString(R.string.noPaper));
-    dlg.setMessage(getString(R.string.noPaperNotice));
-    dlg.setCancelable(false);
-    dlg.setPositiveButton(R.string.sure, new DialogInterface.OnClickListener() {
-      @Override
-      public void onClick(DialogInterface dialogInterface, int i) {
-        ThermalPrinter.stop(ActivityEtatV.this);
-      }
-    });
-    dlg.show();
-  }
-
-
-  private class MyHandler_Bluetooth extends Handler {
-    @Override
-    public void handleMessage(Message msg) {
-      switch (msg.what) {
-        case PRINTVERSION:
-          dialog.dismiss();
-          if (msg.obj.equals("1")) {
-            // textPrintVersion.setText(printVersion);
-          } else {
-            Toast.makeText(ActivityEtatV.this, "Operation failed", Toast.LENGTH_LONG).show();
-          }
-          break;
-        case PRINTCONTENT:
-          new ActivityEtatV.contentPrintThread_Bluetooth().start();
-          break;
-        case CANCELPROMPT:
-          if (progressDialog != null && !ActivityEtatV.this.isFinishing()) {
-            progressDialog.dismiss();
-            progressDialog = null;
-          }
-          break;
-        case OVERHEAT:
-          AlertDialog.Builder overHeatDialog = new AlertDialog.Builder(ActivityEtatV.this);
-          overHeatDialog.setTitle(R.string.operation_result);
-          overHeatDialog.setMessage(getString(R.string.overTemp));
-          overHeatDialog.setPositiveButton(getString(R.string.dialog_comfirm), new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialogInterface, int i) {
-            }
-          });
-          overHeatDialog.show();
-          break;
-        default:
-          Toast.makeText(ActivityEtatV.this, "Print Error!", Toast.LENGTH_LONG).show();
-          break;
-      }
-    }
-  }
-
-
-
-  public void onActivityResult(int requestCode, int resultCode, Intent data) {
-    if(requestCode == BluetoothState.REQUEST_CONNECT_DEVICE) {
-      if(resultCode == Activity.RESULT_OK){
-        bt.connect(data);
-        progressDialog_wait_connecte = ProgressDialog.show(ActivityEtatV.this, getString(R.string.bl_dy1), getString(R.string.printing_wait1));
-      }
-    } else if(requestCode == BluetoothState.REQUEST_ENABLE_BT) {
-      if(resultCode == Activity.RESULT_OK) {
-        bt.setupService();
-        bt.startService(BluetoothState.DEVICE_ANDROID);
-        setup();
-      } else {
-        Toast.makeText(getApplicationContext(), "Bluetooth was not enabled.", Toast.LENGTH_SHORT).show();
-        finish();
-      }
-    }
-
-  }
-
-  //Class thread printing
-  private class contentPrintThread_Bluetooth extends Thread {
-    @Override
-    public void run() {
-      super.run();
-
-      try {
-
-
-        byte[] arrayOfByte1 = { 27, 33, 0 };
-        byte[] format = { 27, 33, 0 };
-
-        // byte[] center =  { 0x1b, 'a', 0x01 }; // center alignment
-
-        // Underline
-        // format[2] = ((byte)(0x80 | arrayOfByte1[2]));
-
-        bt.send(format, true);
-
-        SharedPreferences prefs = getSharedPreferences(PREFS_PRINTER, MODE_PRIVATE);
-        if(prefs.getBoolean("ENTETE_SHOW", false)){
-
-          bt.send(prefs.getString("COMPANY_NAME", ""), true);
-          bt.send(prefs.getString("ACTIVITY_NAME", ""), true);
-          bt.send(prefs.getString("ADRESSE", ""), true);
-          bt.send("--------------------------------", true);
-
-        }
-
-        ////////////////////////////////
-        //bt.send(center, true);
-        format[2] = ((byte) (0x20 | arrayOfByte1[2]));
-        bt.send(format, true);
-
-        ////////////////////////////////
-        //  ThermalPrinter.setBold(true);
-        bt.send(" Statistique des ventes", true);
-        bt.send("------------------------", true);
-
-
-        ////////////////////////////////
-        format[2] = ((byte)(0x0 | arrayOfByte1[2]));
-        bt.send(format, true);
-        bt.send("CLIENT : " + event_selection.getUser(), true);
-        bt.send("DE : " + event_selection.getDate_f(), true);
-        bt.send("VERS : " + event_selection.getDate_t(), true);
-        bt.send("--------------------------------", true);
-
-        ///////////////////////////
-
-        format[2] = ((byte)(0x1 | arrayOfByte1[2]));
-        bt.send(format, true);
-
-        for(int b= 0; b<result_etatzg.size(); b++){
-          if(result_etatzg.get(b).code_parent.equals("1")){
-            if(result_etatzg.get(b).produit.length()>48){
-              result_etatzg.get(b).produit = result_etatzg.get(b).produit.substring(0, 47);
-            }
-
-
-            String quantite = result_etatzg.get(b).quantite;
-            String prix_u = new DecimalFormat("##,##0.00").format(Double.valueOf(result_etatzg.get(b).pv_ht));
-            String total_produit = new DecimalFormat("##,##0.00").format(Double.valueOf(Double.valueOf(result_etatzg.get(b).montant)));
-
-            bt.send(result_etatzg.get(b).produit + "  / " + quantite +" X "+  prix_u +" = " + total_produit, true);
-          }
-        }
-
-        format[2] = ((byte)(0x0 | arrayOfByte1[2]));
-        bt.send(format, true);
-        bt.send("------------------------------------------------", true);
-
-        for(int s = 0; s < result_etatzg.size(); s ++){
-
-          if(result_etatzg.get(s).code_parent.equals("-6")) {
-
-            String walk = "";
-
-            if(result_etatzg.get(s).quantite != null){
-              String total_ht = "TOTAL QTE : " + new DecimalFormat("##,##0").format(Double.valueOf(result_etatzg.get(s).quantite));
-              for (int a = 0; a < 48 - total_ht.length(); a++) {
-                walk = walk + " ";
-              }
-              bt.send(walk + total_ht, true);
-            }
-
-            if(result_etatzg.get(s).montant != null){
-              walk = "";
-              String tva = "MONTANT TOTAL : " + new DecimalFormat("##,##0.00").format(Double.valueOf(result_etatzg.get(s).montant)) + " DA";
-              for (int a = 0; a < 48 - tva.length(); a++) {
-                walk = walk + " ";
-              }
-              bt.send(walk + tva, true);
-            }
-          }
-        }
-
-        bt.send("\n", true);
-
-      } catch (Exception e) {
-        e.printStackTrace();
-        Result = e.toString();
-        handler_Bluetooth.sendMessage(handler_Bluetooth.obtainMessage(PRINTERR, 1, 0, null));
-
-      } finally {
-        handler_Bluetooth.sendMessage(handler_Bluetooth.obtainMessage(CANCELPROMPT, 1, 0, null));
-      }
-    }
-  }
-  public void setup() {
-     /*   Button btnSend = (Button)findViewById(R.id.btnSend);
-        btnSend.setOnClickListener(new View.OnClickListener(){
-            public void onClick(View v){
-                if(etMessage.getText().length() != 0) {
-                    bt.send(etMessage.getText().toString(), true);
-                    etMessage.setText("");
-                }
-            }
-        });
-        */
-  }
 }
