@@ -12,8 +12,8 @@ import android.widget.TextView;
 import com.haozhang.lib.SlantedTextView;
 import com.safesoft.proapp.distribute.R;
 import com.safesoft.proapp.distribute.postData.PostData_Achat1;
-import com.safesoft.proapp.distribute.util.ColorGeneratorModified;
-import com.safesoft.proapp.distribute.util.MyCardView2;
+import com.safesoft.proapp.distribute.utils.ColorGeneratorModified;
+import com.safesoft.proapp.distribute.utils.MyCardView2;
 
 import java.util.List;
 
