@@ -5,6 +5,9 @@ import android.graphics.Color;
 import android.graphics.Typeface;
 import androidx.cardview.widget.CardView;
 import androidx.recyclerview.widget.RecyclerView;
+
+import android.text.SpannableString;
+import android.text.TextUtils;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
@@ -12,10 +15,12 @@ import android.widget.TextView;
 import com.haozhang.lib.SlantedTextView;
 import com.safesoft.proapp.distribute.R;
 import com.safesoft.proapp.distribute.postData.PostData_Inv1;
-import com.safesoft.proapp.distribute.util.ColorGeneratorModified;
-import com.safesoft.proapp.distribute.util.MyCardView2;
+import com.safesoft.proapp.distribute.utils.ColorGeneratorModified;
+import com.safesoft.proapp.distribute.utils.MyCardView2;
 
 import java.util.List;
+
+import cn.nekocode.badge.BadgeDrawable;
 
 /**
  * Created by UK2016 on 02/01/2017.
@@ -24,7 +29,6 @@ import java.util.List;
 public class RecyclerAdapterInv1 extends RecyclerView.Adapter<RecyclerAdapterInv1.MyViewHolder> {
 
   private List<PostData_Inv1> inv1List;
-  private int color = 0;
   private ItemClick itemClick;
   private ItemLongClick itemLongClick;
   private ColorGeneratorModified generator;
@@ -38,6 +42,7 @@ public class RecyclerAdapterInv1 extends RecyclerView.Adapter<RecyclerAdapterInv
     TextView Date_exp;
     TextView nbrProduit;
     TextView Date_inv;
+    TextView Heure_inv;
     CardView cardView;
     SlantedTextView blocage;
 
@@ -50,15 +55,15 @@ public class RecyclerAdapterInv1 extends RecyclerView.Adapter<RecyclerAdapterInv
       Date_exp = (TextView) view.findViewById(R.id.date_exp);
       nbrProduit = (TextView) view.findViewById(R.id.nbr_p);
       Date_inv = (TextView) view.findViewById(R.id.date_inv);
+      Heure_inv = (TextView) view.findViewById(R.id.heure_inv);
       blocage = (SlantedTextView) view.findViewById(R.id.blocage);
+
     }
   }
 
 
   public RecyclerAdapterInv1(Context context, List<PostData_Inv1> itemList) {
     this.inv1List = itemList;
-    if (color == 0)
-      generator = ColorGeneratorModified.MATERIAL;
     mContext = context;
   }
 
@@ -77,29 +82,25 @@ public class RecyclerAdapterInv1 extends RecyclerView.Adapter<RecyclerAdapterInv
   public void onBindViewHolder(final MyViewHolder holder,int position) {
     PostData_Inv1 item = inv1List.get(position);
 
-    holder.NumInv.setTextSize(17);
-    holder.NumInv.setTypeface(null, Typeface.BOLD);
-    holder.NumInv.setText(""+item.num_inv);
+     holder.NumInv.setTextSize(17);
+     holder.NumInv.setTypeface(null, Typeface.BOLD);
+     holder.NumInv.setText(""+item.num_inv);
 
-    holder.NomInventaire.setText(""+item.nom_inv);
-
-            /*
-            if(item.nbr_p == null){
-                item.nbr_p = "0";
-            }
-
-            final BadgeDrawable drawable1 = new BadgeDrawable.Builder()
-                    .type(BadgeDrawable.TYPE_NUMBER)
-                    .badgeColor(0xff303F9F)
-                    .textSize(35)
-                    .number(Integer.valueOf(item.nbr_p))
-                    .build();
+     holder.NomInventaire.setText(""+item.nom_inv);
 
 
-            SpannableString spannableString1 = new SpannableString(TextUtils.concat(drawable1.toSpannable()));
+     final BadgeDrawable drawable1 = new BadgeDrawable.Builder()
+             .type(BadgeDrawable.TYPE_NUMBER)
+             .badgeColor(0xff303F9F)
+             .textSize(35)
+             .number(item.nbr_produit)
+             .build();
 
-            holder.nbrProduit.setText(spannableString1);
- */
+
+     SpannableString spannableString1 = new SpannableString(TextUtils.concat(drawable1.toSpannable()));
+
+     holder.nbrProduit.setText(spannableString1);
+
     holder.Date_inv.setText(item.date_inv + " " + item.heure_inv);
 
     holder.cardView.setOnClickListener(new View.OnClickListener() {
@@ -117,31 +118,31 @@ public class RecyclerAdapterInv1 extends RecyclerView.Adapter<RecyclerAdapterInv
       }
     });
 
-    if(item.is_sent == 1){
-      holder.blocage.setText("Exporté")
+
+    if(item.blocage.equals("F")){
+      holder.blocage.setText("Validé")
               .setTextColor(Color.WHITE)
               .setSlantedBackgroundColor(Color.GREEN)
               .setTextSize(21)
-              .setSlantedLength(50)
+              .setSlantedLength(80)
               .setMode(SlantedTextView.MODE_RIGHT_BOTTOM);
-
-      holder.Date_exp.setText(item.date_export_inv);
     }else {
       holder.blocage.setText("En attente")
               .setTextColor(Color.WHITE)
               .setSlantedBackgroundColor(Color.RED)
               .setTextSize(21)
-              .setSlantedLength(50)
+              .setSlantedLength(80)
               .setMode(SlantedTextView.MODE_RIGHT_BOTTOM);
+    }
+
+    if(item.is_sent == 1){
+      holder.Date_exp.setText(item.date_export_inv);
+    }else {
       holder.Date_exp.setText("Pas encore exporté");
     }
 
-
-    if (color == 0){
-      if (generator!=null)
-        color = generator.getColor(inv1List.get(position).num_inv);
-    }
   }
+
 
   @Override
   public int getItemCount() {
@@ -152,13 +153,16 @@ public class RecyclerAdapterInv1 extends RecyclerView.Adapter<RecyclerAdapterInv
     void onClick(View v, int position);
   }
 
+
   public interface ItemLongClick{
     void onLongClick(View v, int position);
   }
+
 
   public void refresh(List<PostData_Inv1> new_itemList){
     inv1List.clear();
     inv1List.addAll(new_itemList);
     notifyDataSetChanged();
   }
+
 }
