@@ -8,20 +8,24 @@ public class PostData_Carnet_c  {
     public String code_client;
     public String carnet_date;
     public String carnet_heure;
-    public String carnet_achats="0.00";
-    public String carnet_versement;
+    public Double carnet_achats=0.00;
+    public Double carnet_versement;
     public String carnet_source = "SITUATION-CLIENT";
     public String carnet_num_bon = null;
     public String carnet_mode_rg = "ESPECE";
     public String carnet_remarque;
     public String carnet_utilisateur = "TERMINAL_MOBILE";
     public String exportation;
+    public String client;
     public String adresse;
     public String tel;
     public String longitude;
     public String latitude;
     public String rc;
     public String ifiscal;
+    public String ai;
+    public String nis;
+    public String mode_tarif;
 
 
 }
