@@ -4,7 +4,6 @@ package com.safesoft.proapp.distribute.postData;
  * Created by UK2015 on 31/08/2016.
  */
 public class PostData_Client {
-    public String client_id = " ";
     public String code_client = " ";
     public String client = " ";
     public String tel = " ";
@@ -12,12 +11,13 @@ public class PostData_Client {
     public Double latitude = 0.0;
     public Double longitude = 0.0;
     public String mode_tarif = " ";
-    public String achat_montant = "0.00";
-    public String verser_montant = "0.00";
-    public String solde_montant = "0.00";
+    public Double achat_montant = 0.00;
+    public Double verser_montant = 0.00;
+    public Double solde_montant = 0.00;
 
     public String rc;
     public String ifiscal;
-    public String rib;
+    public String nis;
+    public String ai;
     public Double credit_limit = 0.00;
 }
