@@ -4,14 +4,20 @@ package com.safesoft.proapp.distribute.postData;
  * Created by UK2015 on 24/08/2016.
  */
 public class PostData_Inv2 {
+    public Integer recordid;
     public String reference;
     public String produit;
     public String codebarre;
-    public String pa_ht;
-    public String quantity_old;
-    public String tva;
-    public String quantity_new;
+    public Double pa_ht;
+    public Double qte_theorique;
+    public Double qte_physique;
+    public Double tva;
     public String num_inv;
+    public String code_depot;
+
+    public Double nbr_colis;
+    public Double colissage;
+    public Double vrac;
 
     public Boolean exist;
 }
